+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented here.
 
 ### Changed
 - `STRINGIFY_INDEX_VALUES` is `False` by default (index `.name` will still be a string, but values will keep their original type)
+- Added comms listener for renaming (SQL cell) dataframes.
 
 ## `1.2.0`
 _2022-08-21_